change type

+++ nc-150/binary-search/koko-eating-bananas/1.c
@@ -6,7 +6,7 @@
 int min_eating_speed(int *piles, int n, int h) {
   int k = 1;
   for (;;) {
-    int t = 0;
+    long long t = 0;
     for (int i = 0; i < n; i++) {
       t += ceil((double)piles[i] / k);
     }
